#1942 moved prodile to separate util
