+++ controller/common/profile/server-old-hash.js
@@ -1,0 +1,15 @@
+'use strict';
+
+var oldClientHash = require('mano-auth/utils/client-hash')
+  , profileSubmit = require('./server').profile.submit;
+
+exports.profile = {
+	submit: function (normalizedData, data) {
+		if (normalizedData.password || normalizedData['password-new']) {
+			normalizedData.password = oldClientHash(this.user.email, normalizedData.password);
+			normalizedData['password-new'] =
+				oldClientHash(this.user.email, normalizedData['password-new']);
+		}
+		return profileSubmit.apply(this, arguments);
+	}
+};
