+++ controller/user/index.js
@@ -1,24 +1,15 @@
 'use strict';
 
-var changeOwnPassword      = require('mano-auth/controller/change-own-password').validate
-  , validate               = require('mano/utils/validate')
-  , env                    = require('mano').env
+var env                    = require('mano').env
   , assign                 = require('es5-ext/object/assign')
   , externalAuthentication = (env && env.externalAuthentication) || {};
 
-assign(exports, require('../common/managed-profile'), require('../common/request-create-account'));
+assign(exports, require('../common/managed-profile'),
+	require('../common/request-create-account'), require('../common/profile/index'));
 
 // Profile
-if (!externalAuthentication.profilePage) {
-	exports.profile = {
-		validate: function (data) {
-			if (data.password || data['password-new']) return changeOwnPassword.call(this, data);
-			delete data.password;
-			delete data['password-new'];
-			delete data['password-repeat'];
-			return validate.call(this, data, { partial: true });
-		}
-	};
+if (externalAuthentication.profilePage) {
+	delete exports.profile;
 }
 
 // Role switch
