+++ controller/user/server-old-hash.js
@@ -3,21 +3,12 @@
 var assign                 = require('es5-ext/object/assign')
   , oldClientHash          = require('mano-auth/utils/client-hash')
   , env                    = require('mano').env
-  , profileSubmit          = require('./server').profile.submit
   , externalAuthentication = (env && env.externalAuthentication) || {};
 
 assign(exports, require('./server'),
-	require('../demo-user/server')({ oldClientHash: oldClientHash }));
+	require('../demo-user/server')({ oldClientHash: oldClientHash }),
+	require('../common/profile/server-old-hash'));
 
-if (!externalAuthentication.profilePage) {
-	exports.profile = {
-		submit: function (normalizedData, data) {
-			if (normalizedData.password || normalizedData['password-new']) {
-				normalizedData.password = oldClientHash(this.user.email, normalizedData.password);
-				normalizedData['password-new'] =
-					oldClientHash(this.user.email, normalizedData['password-new']);
-			}
-			return profileSubmit.apply(this, arguments);
-		}
-	};
+if (externalAuthentication.profilePage) {
+	delete exports.profile;
 }
